feat: maximum theoretical margin b3 and spot accepted as collateral in b3 added to ingestion
