+++ stpstone/analytics/b3/search_by_trading.py
@@ -29,455 +29,6 @@ class TradingFilesB3:
         self.str_fillna_ts = str_fillna_ts
         self.str_fmt_dt = str_fmt_dt
 
-    def options_b3(self, float_value_error=-1000.0, bl_debug=False,
-                   float_days_year=365.0):
-        """
-        DOCSTRING:
-        INPUTS:
-        OUTPUTS:
-        """
-        # setting variables
-        list_ser = list()
-        # raw maximum theoretical margins
-        list_rows_mtm = self.carga_mtm_b3
-        # b3 financial indicators
-        df_fin_ind = MDB3().financial_indicators_b3
-        # risk-free rate
-        r = df_fin_ind[df_fin_ind[YAML_B3['financial_indicators']['col_si_code']] \
-                       == YAML_B3['financial_indicators']['int_si_code_selic']][
-                           YAML_B3['financial_indicators']['col_rate']
-                       ].tolist()[0]
-        if bl_debug == True:
-            print('RISK-FREE RATE: {}'.format(r))
-        # options traded in b3 exchange - in zip file
-        url = YAML_B3['options_traded_b3'][
-            'url'].format(DatesBR().sub_working_days(DatesBR().curr_date,
-                                                     self.int_wd_bef).strftime('%y%m%d'))
-        # carga para a memória margens teóricas máximas b3
-        zipfile = RemoteFiles().get_zip_from_web_in_memory(
-            url, bl_io_interpreting=YAML_B3[
-                'options_traded_b3']['bl_io_interpreting'],
-            bl_verify=YAML_B3['options_traded_b3']['bl_verify']
-        )
-        # retorno é um novo zip com dois xmls
-        zipfile = ZipFile(zipfile)
-        # iterate through the files in the zip
-        for file_info in zipfile.infolist():
-            #   opennig just xml files
-            if file_info.filename.endswith('.xml'):
-                xml_file = zipfile.open(file_info)
-                xml_file = xml_file.read()
-            else:
-                continue
-            #   soup parser to xml trading options
-            soup_xml = XMLFiles().xml_memory_parser(xml_file)
-            #   looping through tags of BOV & BMF markets
-            for key, value in YAML_B3['options_traded_b3']['tags'].items():
-                #   finding key of interest within soup
-                for soup_content in soup_xml.find_all(key):
-                    #   reseting variables
-                    dict_ = dict()
-                    bl_break_opt_price = False
-                    bl_break_spt_price = False
-                    #   adding all tags availabe to the dictionary
-                    for tag in value:
-                        dict_[tag] = soup_content.find(tag).get_text()
-                        if bl_debug == True:
-                            print(soup_content.find(tag).get_text())
-                    if bl_debug == True:
-                        print('ISIN: {}'.format(soup_content.find(YAML_B3['options_traded_b3'][
-                            'keys']['isin']).get_text()))
-                        print('TICKER: {}'.format(dict_[YAML_B3['options_traded_b3']['keys'][
-                            'ticker']]))
-                    #   spot ticker
-                    str_isin = soup_content.find(YAML_B3['options_traded_b3']['keys']['isin']).get_text()
-                    dict_[YAML_B3['options_traded_b3']['keys']['spot']] = \
-                        str_isin[2:6] + str_isin[6].replace('9', '11')
-                    if bl_debug == True:
-                        print('SPOT TICKER: {}'.format(dict_[YAML_B3['options_traded_b3'][
-                            'keys']['spot']]))
-                    #   calendar days to expiration
-                    dict_[YAML_B3['options_traded_b3']['keys']['cd_exp']] = \
-                        DatesBR().delta_calendar_days(
-                            DatesBR().sub_working_days(
-                                DatesBR().curr_date, abs(-self.int_wd_bef-1)),
-                                DatesBR().str_date_to_datetime(str(dict_[
-                                        YAML_B3['options_traded_b3']['keys']['xpr_dt']
-                                    ]), YAML_B3['options_traded_b3']['dt_input_format']
-                            )
-                        )
-                    if bl_debug == True:
-                        print('CALENDAR DAYS TO EXPIRATION: {}'.format(
-                                DatesBR().delta_calendar_days(
-                                DatesBR().sub_working_days(
-                                    DatesBR().curr_date, abs(-self.int_wd_bef-1)),
-                                    DatesBR().str_date_to_datetime(str(dict_[
-                                            YAML_B3['options_traded_b3']['keys']['xpr_dt']
-                                        ]), YAML_B3['options_traded_b3']['dt_input_format']
-                                    )
-                                )
-                            )
-                        )
-                    #   closing price of spot and option
-                    while \
-                        (bl_break_spt_price == False) \
-                        and (bl_break_opt_price == False):
-                        for i, _ in enumerate(list_rows_mtm):
-                            if i < len(list_rows_mtm) - 3:
-                                #   option close
-                                if list_rows_mtm[i][0] == YAML_B3['margens_teoricas_maximas_b3'][
-                                    'key_instrumentos'] \
-                                    and list_rows_mtm[i][4] != '0' \
-                                    and list_rows_mtm[i][5] == dict_[
-                                        YAML_B3['options_traded_b3']['keys']['ticker']]:
-                                    dict_[YAML_B3['options_traded_b3']['keys']['close_opt']] = \
-                                        float(list_rows_mtm[i][4].replace(",", "."))
-                                    bl_break_opt_price = True
-                                    if bl_debug == True:
-                                        print('OPT CLOSE: {}'.format(float(
-                                            list_rows_mtm[i][4].replace(",", "."))))
-                                #   spot close
-                                if list_rows_mtm[i][0] == YAML_B3['margens_teoricas_maximas_b3'][
-                                    'key_instrumentos'] \
-                                    and list_rows_mtm[i][4] != '0' \
-                                    and list_rows_mtm[i][5] == dict_[YAML_B3['options_traded_b3'][
-                                        'keys']['spot']]:
-                                    dict_[YAML_B3['options_traded_b3']['keys']['close_spot']] = \
-                                        float(list_rows_mtm[i][4].replace(",", "."))
-                                    bl_break_spt_price = True
-                                    if bl_debug == True:
-                                        print('SPOT CLOSE: {}'.format(float(
-                                            list_rows_mtm[i][4].replace(",", "."))))
-                    #   risk-free rate
-                    dict_[YAML_B3['options_traded_b3']['keys']['risk_free']] = r
-                    #   days to maturity ratio
-                    dict_[YAML_B3['options_traded_b3']['keys']['days_maturity_ratio']] = \
-                        float(dict_[YAML_B3['options_traded_b3']['keys']['cd_exp']]) \
-                        / float_days_year
-                    #   implied volatility
-                    if bl_debug == True:
-                        print('*** INPUTS VOL IMP ***')
-                        print(
-                            float(dict_[YAML_B3['options_traded_b3']['keys']['close_spot']]),
-                            float(dict_[YAML_B3['options_traded_b3']['keys']['k']]),
-                            float(r),
-                            float(dict_[YAML_B3['options_traded_b3']['keys'][
-                                'days_maturity_ratio']]),
-                            float(0.5),
-                            float(0),
-                            float(r),
-                            float(dict_[YAML_B3['options_traded_b3']['keys']['close_opt']]),
-                            str(dict_[YAML_B3['options_traded_b3']['keys'][
-                                'opt_type']].lower().replace('putt', 'put'))
-                        )
-                    try:
-                        dict_[YAML_B3['options_traded_b3']['keys']['imp_vol']] = \
-                            EuropeanOptions().implied_volatility(
-                                float(dict_[YAML_B3['options_traded_b3']['keys']['close_spot']]),
-                                float(dict_[YAML_B3['options_traded_b3']['keys']['k']]),
-                                float(r),
-                                float(float(dict_[YAML_B3['options_traded_b3']['keys']['cd_exp']]) \
-                                    / 365.0),
-                                float(0.5),
-                                float(0),
-                                float(r),
-                                float(dict_[YAML_B3['options_traded_b3']['keys']['close_opt']]),
-                                str(dict_[YAML_B3['options_traded_b3']['keys']['opt_type']
-                                        ].lower().replace('putt', 'put')
-                                )
-                            )
-                    except KeyError:
-                        dict_[YAML_B3['options_traded_b3']['keys']['imp_vol']] = float_value_error
-                    #   delta
-                    try:
-                        dict_[YAML_B3['options_traded_b3']['keys']['delta']] = \
-                            EuropeanOptions().delta(
-                                float(dict_[YAML_B3['options_traded_b3']['keys']['close_spot']]),
-                                float(dict_[YAML_B3['options_traded_b3']['keys']['k']]),
-                                float(r),
-                                float(float(dict_[YAML_B3['options_traded_b3']['keys']['cd_exp']]) \
-                                    / 365.0),
-                                dict_[YAML_B3['options_traded_b3']['keys']['imp_vol']],
-                                float(0),
-                                float(r),
-                                str(dict_[YAML_B3['options_traded_b3']['keys']['opt_type']
-                                    ].lower().replace('putt', 'put')
-                                )
-                            )
-                    except KeyError:
-                        dict_[YAML_B3['options_traded_b3']['keys']['delta']] = float_value_error
-                    if bl_debug == True:
-                        print('DELTA: {}'.format(dict_[YAML_B3['options_traded_b3']['keys'][
-                            'delta']]))
-                        print('IV: {}'.format(dict_[YAML_B3['options_traded_b3']['keys'][
-                            'imp_vol']]))
-                    #   appendig dictionary to list
-                    list_ser.append(dict_)
-                    if bl_debug == True:
-                        print(dict_)
-        # uploading list of dictionaries to pandas
-        df_opt_in = pd.DataFrame(list_ser)
-        # removing duplicates
-        df_opt_in.drop_duplicates(inplace=True)
-        # adding logging
-        df_opt_in = DBLogs().audit_log(
-            df_opt_in,
-            url,
-            DatesBR().utc_from_dt(
-                DatesBR().sub_working_days(DatesBR().curr_date, self.int_wd_bef)
-            )
-        )
-        # returning dataframe
-        return df_opt_in
-
-    @property
-    def carga_mtm_b3(self):
-        """
-        DOCSTRING:
-        INPUTS:
-        OUTPUTS:
-        """
-        # url de exportação do arquivo de margens teóricas máximas b3
-        url_mtm = YAML_B3['margens_teoricas_maximas_b3'][
-            'url'].format(DatesBR().sub_working_days(DatesBR().curr_date,
-                                                     self.int_wd_bef).strftime('%y%m%d'))
-        # carga para a memória margens teóricas máximas b3
-        zipfile = RemoteFiles().get_zip_from_web_in_memory(
-            url_mtm, bl_io_interpreting=YAML_B3[
-                'margens_teoricas_maximas_b3']['bl_io_interpreting'],
-            bl_verify=YAML_B3['margens_teoricas_maximas_b3']['bl_verify'])
-        # retorno é um novo zip com um xlsx
-        zipfile = ZipFile(zipfile)
-        # coletando nome do arquivo txt contido dentro do zip
-        list_zip_files = zipfile.namelist()
-        file_name = list_zip_files.pop()
-        # extraindo arquivo com fatores de risco primitivos b3 - txt
-        name_file_extract = zipfile.open(file_name)
-        # lendo csv com margens teóricas máximas b3
-        list_rows_mtm = [
-            str(str_).replace(r'\n', '').replace(
-                "b'", '').replace("'", '').split(';')
-            for str_ in name_file_extract.readlines() if len(
-                str(str_).replace(r'\n', '').replace("b'", '').replace(
-                    "'", '').split(';')) > 1
-        ]
-        list_rows_mtm = list_rows_mtm[1:]
-        # retornando lista de linhas do csv corrente
-        return list_rows_mtm
-
-    @property
-    def mtm_compra_venda(self, den_desagios=None,
-                         key_stocks='stocks', key_funds='funds', key_etfs='etfs', key_bdrs='bdrs'):
-        """
-        DOCSTRING:
-        INPUTS:
-        OUTPUTS
-        """
-        # setting variables
-        list_num_desagios_compra = list()
-        list_num_desagios_venda = list()
-        list_exportacao = list()
-        # carga de margens teóricas máximas crua
-        list_rows_mtm = self.carga_mtm_b3
-        # carga de ativos bovespa e bmf por tipo
-        dict_ativos_bov_bmf_tipo = UP2DATAB3().security_category_name
-        # selecionando apenas lista de ativos de interesse - ações, fiis, etfs, e bdrs
-        list_ativos_bov_vista = ListHandler().extend_lists(
-            dict_ativos_bov_bmf_tipo[key_stocks], dict_ativos_bov_bmf_tipo[key_funds],
-            dict_ativos_bov_bmf_tipo[key_etfs], dict_ativos_bov_bmf_tipo[key_bdrs])
-        # removendo eventuais duplicatas
-        list_ativos_bov_vista = ListHandler().remove_duplicates(list_ativos_bov_vista)
-        # alocando margens teóricas máximas em lista de dicionários
-        for i, _ in enumerate(list_rows_mtm):
-            if i < len(list_rows_mtm) - 3:
-                #   setando variáveis de passagem
-                dict_exportacao = dict()
-                #   margem B3 requerida para passar posicionado no papel
-                if list_rows_mtm[i][0] == YAML_B3['margens_teoricas_maximas_b3'][
-                    'key_instrumentos'] \
-                    and list_rows_mtm[i][4] != '0' \
-                    and (len(list_rows_mtm[i][5]) == 5 or
-                         len(list_rows_mtm[i][5]) == 6):
-                    den_desagios = float(
-                        list_rows_mtm[i][4].replace(",", "."))
-                #   numerador de preço com choque para cálculo de deságio por cenário
-                #       dias úteis considerados: 1 a 10 dias úteis
-                #       garantias exigidas calculadas de compra e venda
-                for dias_uteis_choque in range(1, 11):
-                    try:
-                        if int(list_rows_mtm[i + dias_uteis_choque][1]) == \
-                            dias_uteis_choque \
-                                and list_rows_mtm[i + dias_uteis_choque][2] != '':
-                            list_num_desagios_compra.append(float(
-                                list_rows_mtm[i + dias_uteis_choque][2].replace(",", ".")))
-                        if int(list_rows_mtm[i + dias_uteis_choque][
-                            1]) == dias_uteis_choque and list_rows_mtm[
-                                i + dias_uteis_choque][3] != '':
-                            list_num_desagios_venda.append(float(
-                                list_rows_mtm[i + dias_uteis_choque][3].replace(",", ".")))
-                    except:
-                        pass
-                # executar o loop para listas vazias
-                if (list_num_desagios_compra != []) and (list_num_desagios_venda != []):
-                    # denominador com preço spot no fechamento para cálculo do deságio
-                    #   B3 por cenário
-                    if (list_rows_mtm[i][0] == YAML_B3['margens_teoricas_maximas_b3'][
-                        'key_instrumentos']) and (list_rows_mtm[i][4] != '0') \
-                            and all([type(x) == float for x in
-                                     list_num_desagios_compra]) and (type(
-                            den_desagios) == float):
-                        # setando dicionário para o ticker de interesse
-                        dict_exportacao[YAML_B3['margens_teoricas_maximas_b3'][
-                            'col_incluir_ticker']] = list_rows_mtm[i][5]
-                        # definindo preço de fechamento de referência
-                        dict_exportacao['PRECO_FECHAMENTO'] = den_desagios
-                        # definindo deságios para cenários de choque para o período de
-                        #   interesse
-                        for dias_uteis_choque in range(1, 11):
-                            # deságio de compra/venda para ativo corrente negociado em mercado
-                            #   secundário da b3
-                            if list_rows_mtm[i][5] in list_ativos_bov_vista:
-                                dict_exportacao[YAML_B3['margens_teoricas_maximas_b3'][
-                                    'col_incluir_margem_b3_compra_d_mais_n'] + str(
-                                    dias_uteis_choque)] = abs(den_desagios * (
-                                        1 - list_num_desagios_compra[
-                                            dias_uteis_choque - 1] / den_desagios))
-                                dict_exportacao[YAML_B3['margens_teoricas_maximas_b3'][
-                                    'col_incluir_desagio_b3_compra_d_mais_n'] + str(
-                                    dias_uteis_choque)] = min([1, abs(1 - list_num_desagios_compra[
-                                        dias_uteis_choque - 1] / den_desagios)])
-                                dict_exportacao[YAML_B3['margens_teoricas_maximas_b3'][
-                                    'col_incluir_margem_b3_venda_d_mais_n'] + str(
-                                    dias_uteis_choque)] = \
-                                    abs(den_desagios * (
-                                        1 - list_num_desagios_venda[
-                                            dias_uteis_choque - 1] / den_desagios))
-                                dict_exportacao[YAML_B3['margens_teoricas_maximas_b3'][
-                                    'col_incluir_desagio_b3_venda_d_mais_n'] + str(
-                                    dias_uteis_choque)] = \
-                                    min([1, abs(1 - list_num_desagios_venda[
-                                        dias_uteis_choque - 1] / den_desagios)])
-                            else:
-                                dict_exportacao[YAML_B3['margens_teoricas_maximas_b3'][
-                                    'col_incluir_margem_b3_compra_d_mais_n'] + str(
-                                    dias_uteis_choque)] = abs(
-                                    -list_num_desagios_compra[
-                                        dias_uteis_choque - 1])
-                                dict_exportacao[YAML_B3['margens_teoricas_maximas_b3'][
-                                    'col_incluir_desagio_b3_compra_d_mais_n'] + str(
-                                    dias_uteis_choque)] = \
-                                    min([1, abs(list_num_desagios_compra[
-                                        dias_uteis_choque - 1] / den_desagios - 1)])
-                                dict_exportacao[YAML_B3['margens_teoricas_maximas_b3'][
-                                    'col_incluir_margem_b3_venda_d_mais_n'] + str(
-                                    dias_uteis_choque)] = \
-                                    abs(-list_num_desagios_venda[dias_uteis_choque - 1])
-                                dict_exportacao[YAML_B3['margens_teoricas_maximas_b3'][
-                                    'col_incluir_desagio_b3_venda_d_mais_n'] + str(
-                                    dias_uteis_choque)] = \
-                                    min([1, abs(list_num_desagios_venda[
-                                        dias_uteis_choque - 1] / den_desagios - 1)])
-                    # appendando à lista de exportação
-                    list_exportacao.append(dict_exportacao)
-                    # ressetiando variável de passagem
-                    list_num_desagios_compra = list()
-                    list_num_desagios_venda = list()
-        # appendando lista de exportação a um dataframe
-        df_mtm_b3 = pd.DataFrame(list_exportacao)
-        # drop not available rows
-        df_mtm_b3.dropna(inplace=True)
-        # sort tickers in alphabetic order
-        df_mtm_b3.sort_values([YAML_B3['margens_teoricas_maximas_b3'][
-            'col_incluir_ticker']], ascending=[True], inplace=True)
-        # adding logging
-        df_mtm_b3 = DBLogs().audit_log(
-            df_mtm_b3,
-            YAML_B3['margens_teoricas_maximas_b3'][
-                'url'].format(DatesBR().sub_working_days(
-                    DatesBR().curr_date, self.int_wd_bef).strftime('%y%m%d')),
-            DatesBR().utc_from_dt(
-                DatesBR().sub_working_days(DatesBR().curr_date, self.int_wd_bef)
-            )
-        )
-        # retornando deságios b3 de interesse
-        return df_mtm_b3
-
-    @property
-    def collateral_acc_spot_bov_b3(self, dict_repalce_str_erros={
-            'ETF': ''}, col_ticker='TICKER', col_isin='ISIN', col_limite_qtds='LIMITE_QUANTIDADES',
-            col_mes_arquivo='MES_ARQUIVO', key_limite_qtds='Limite (quantidade)'):
-        """
-        DOCSTRING:
-        INPUTS:
-        OUTPUTS:
-        """
-        # setting variables
-        list_exportacao = list()
-        # importing to memory html, in order to catch the url of the collateral accepted by b3,
-        #   regarding spot bov
-        html_parser_acoes_units_etfs_aceitos_garantia_b3 = HtmlHandler().html_lxml_parser(
-            YAML_B3['collateral_acc_spot_bov_b3']['url'], bl_verify=False)
-        url_acoes_units_etfs_aceitos_garantia_b3 = HtmlHandler().html_lxml_xpath(
-            html_parser_acoes_units_etfs_aceitos_garantia_b3,
-            YAML_B3['collateral_acc_spot_bov_b3'][
-                'xpath_url_acoes_units_etfs_aceitos_garantia_b3'])[0]
-        url_acoes_units_etfs_aceitos_garantia_b3 = StrHandler().get_string_after_substr(
-            url_acoes_units_etfs_aceitos_garantia_b3, YAML_B3['collateral_acc_spot_bov_b3'][
-            'data_files_collateral_acc_spot_bov_b3'])
-        # creating complete url
-        url_acoes_units_etfs_aceitos_garantia_b3 = \
-            YAML_B3['collateral_acc_spot_bov_b3']['url_raiz_documentos_b3'] \
-            + YAML_B3['collateral_acc_spot_bov_b3'][
-            'data_files_collateral_acc_spot_bov_b3'] + \
-            url_acoes_units_etfs_aceitos_garantia_b3
-        # openning zip in memory
-        list_unzipped_files = RemoteFiles().get_zip_from_web_in_memory(
-            url_acoes_units_etfs_aceitos_garantia_b3, bl_verify=False)
-        if str(type(list_unzipped_files)) != "<class 'list'>":
-            list_unzipped_files = [list_unzipped_files]
-        # looping within files in zip
-        for file_unz in list_unzipped_files[1:]:
-            #   creating columns month-file
-            str_mes_arquivo_garantias_b3_corrente = StrHandler().get_between(
-                str(file_unz), '_', '.xlsx').replace(' ', '').lower()
-            #   looping within sheet names
-            for nome_plan_ativa in [
-                StrHandler().latin_characters(
-                    YAML_B3['collateral_acc_spot_bov_b3']['plan_acoes_bdrs_etfs_units']),
-                YAML_B3['collateral_acc_spot_bov_b3']['plan_adrs']
-            ]:
-                #   defining column ticker
-                if nome_plan_ativa == 'ADR':
-                    col_cod_ticker = 'Código'
-                else:
-                    col_cod_ticker = 'Código'
-                #   openning dataframe in memory
-                reader = pd.read_excel(
-                    file_unz, sheet_name=nome_plan_ativa)
-                df_coll_acc_spot_bov = pd.DataFrame(reader)
-                #   creating serialized list with spot bov collateral accepted
-                for _, row in df_coll_acc_spot_bov.iterrows():
-                    list_exportacao.append({
-                        col_mes_arquivo: str_mes_arquivo_garantias_b3_corrente.upper(),
-                        col_ticker: row[col_cod_ticker],
-                        col_isin: row[col_isin],
-                        col_limite_qtds: row[key_limite_qtds]
-                    })
-        # creating dataframe
-        df_coll_acc_spot_bov = pd.DataFrame(list_exportacao)
-        # removing duplicates
-        df_coll_acc_spot_bov[col_ticker] = [StrHandler().replace_all(
-            str(x), dict_repalce_str_erros) for x in df_coll_acc_spot_bov[
-                col_ticker].tolist()]
-        # adding logging
-        df_fpr_b3 = DBLogs().audit_log(
-            df_fpr_b3,
-            url_acoes_units_etfs_aceitos_garantia_b3,
-            DatesBR().utc_from_dt(DatesBR().curr_date)
-        )
-        # returning dataframe
-        return df_coll_acc_spot_bov
-
     def cenarios_tipo_curva(self, list_cenarios_tipo_curva=None):
         """
         DOCSTRING:
